refactor(grow): extract module-level helpers to _helpers.py
Move GrowStageError, _get_prompts_path, _format_structural_feedback,
TypeVar T, and log to grow/_helpers.py. stage.py imports from _helpers.
