+++ src/questfoundry/pipeline/stages/grow/stage.py
@@ -22,8 +22,8 @@ from __future__ import annotations
 from collections.abc import Awaitable, Callable
 from dataclasses import dataclass
 from functools import partial
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, TypeVar
+from pathlib import Path  # noqa: TC003 - used at runtime
+from typing import TYPE_CHECKING, Any
 
 from langchain_core.messages import HumanMessage, SystemMessage
 from pydantic import BaseModel, ValidationError
@@ -47,10 +47,16 @@ from questfoundry.graph.context_compact import (
 from questfoundry.graph.graph import Graph
 from questfoundry.graph.mutations import GrowMutationError, GrowValidationError
 from questfoundry.models.grow import GrowPhaseResult, GrowResult
-from questfoundry.observability.logging import get_logger
 from questfoundry.observability.tracing import traceable
 from questfoundry.pipeline.batching import batch_llm_calls
 from questfoundry.pipeline.gates import AutoApprovePhaseGate
+from questfoundry.pipeline.stages.grow._helpers import (
+    GrowStageError,
+    T,
+    _format_structural_feedback,
+    _get_prompts_path,
+    log,
+)
 from questfoundry.prompts.compiler import safe_format
 from questfoundry.providers.structured_output import (
     unwrap_structured_result,
@@ -72,63 +78,6 @@ if TYPE_CHECKING:
         PhaseProgressFn,
         UserInputFn,
     )
-
-T = TypeVar("T", bound=BaseModel)
-
-
-def _get_prompts_path() -> Path:
-    """Get the prompts directory path.
-
-    Returns prompts from package first, then falls back to project root.
-    """
-    pkg_path = Path(__file__).parents[5] / "prompts"
-    if pkg_path.exists():
-        return pkg_path
-    return Path.cwd() / "prompts"
-
-
-log = get_logger(__name__)
-
-
-class GrowStageError(ValueError):
-    """Error raised when GROW stage cannot proceed."""
-
-    pass
-
-
-def _format_structural_feedback(errors: list[GrowValidationError]) -> str:
-    """Format structural compatibility errors as targeted LLM feedback.
-
-    Categorizes errors by type and produces a correction message that
-    tells the LLM exactly what went wrong, enabling targeted repair.
-    """
-    dilemma_substrings = ("same dilemma", "span only", "at most 1 beat per dilemma")
-    same_dilemma_count = sum(
-        1 for e in errors if any(sub in e.issue.lower() for sub in dilemma_substrings)
-    )
-
-    lines = ["\n\n## CORRECTION REQUIRED"]
-    lines.append(f"Your previous {len(errors)} error(s) caused ALL intersections to be REJECTED.")
-
-    if same_dilemma_count > 0:
-        lines.append(
-            "PROBLEM: You grouped beats from the SAME dilemma together. "
-            "Each intersection MUST combine beats from DIFFERENT dilemmas. "
-            "Each candidate group already contains beats from different dilemmas -- "
-            "select beats from WITHIN a group, not across groups."
-        )
-
-    # Show up to 3 specific errors for clarity
-    max_shown = 3
-    for err in errors[:max_shown]:
-        lines.append(f"  - {err.issue}")
-    if len(errors) > max_shown:
-        lines.append(f"  ... and {len(errors) - max_shown} more error(s)")
-
-    lines.append(
-        "Try again. Use the beat IDs from WITHIN each candidate group to form intersections."
-    )
-    return "\n".join(lines)
 
 
 class GrowStage:
